feat: data pipeline
Created and finished data pipeline for AZDIAS and Customer data. I
intend to use this same pipeline for the mailer data, as it should be
the same. There are 4 major steps to the pipeline:

* Identify Columns that will be dropped due to missingness
* Clean data. This is the bulk of the pipeline
* Scale data. A scaler is created from the AZDIAS data then used on both
AZDIAS and Customer Data
* Impute data. Similar set up to scaled data

This breakdown of steps should allow me to work on any piece independent
from the other pieces when it comes time to fit my models, if there are
any adjustments that need to be made